backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
chromadb==0.5.0
pydantic==2.7.1
python-dotenv==1.0.1
pylint==3.2.0
